format fixed, tests, README
